# Optimize decimal formatting of 128-bit integers